requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.2
attrs==22.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==22.10.0
CacheControl==0.12.11
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
commonmark==0.9.1
Cython==0.29.32
distlib==0.3.6

Django==4.1.3
django-bootstrap4==22.2
django-debug-toolbar==3.7.0
docopt==0.6.2
docutils==0.19
future==0.18.2
Pillow==9.3.0
six==1.16.0
pyparsing==3.0.9
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.12
validate-pyproject==0.10.1

djangorestframework==3.14.0
pymemcache==4.0.0

PyMySQL~=1.0.2